Move events to common
