+++ dbt/adapters/bigquery/connections.py
@@ -35,9 +35,9 @@ from dbt.exceptions import (
     DbtProfileError,
 )
 from dbt.adapters.base import BaseConnectionManager, Credentials
-from dbt.events import AdapterLogger
-from dbt.events.functions import fire_event
-from dbt.events.types import SQLQuery
+from dbt.common.events import AdapterLogger
+from dbt.common.events.functions import fire_event
+from dbt.common.events.types import SQLQuery
 from dbt.version import __version__ as dbt_version
 
 from dbt.dataclass_schema import ExtensibleDbtClassMixin, StrEnum
