+++ dbt/adapters/bigquery/gcloud.py
@@ -1,4 +1,4 @@
-from dbt.events import AdapterLogger
+from dbt.common.events import AdapterLogger
 import dbt.exceptions
 from dbt.clients.system import run_cmd
